# Fix cut off full width glyphs in last column