… CHANGELOG.md
… alacritty_terminal/src/term/cell.rs
… alacritty_terminal/src/term/mod.rs