@chrisduerr 2019-12-15
Seems like the performance of this PR is actually better than before, after some minor tweaks.

Waiting for @kchibisov before merge, since he wanted to take a look..