@kchibisov 2019-12-16
There's a strange bug in bash. I've created a url with these symbols, and then start moving it to the new line. The space, that you can see wasn't added by me. and in fact if I move back to an original line everything will be the same. You can also see that a cursor is broken.

![picture](https://user-images.githubusercontent.com/27620401/70907294-8ec88a80-2019-11ea-994d-236a1f223c11.png)

![picture](https://user-images.githubusercontent.com/27620401/70907387-c8999100-2019-11ea-995c-8b1bf3966950.png)