@kchibisov 2019-12-16
Moving double width chars(test this with URL) is inconsistent a bit. Sometimes you move entire chars and adding a space, sometimes they moved only half way, resulting a broken highlighting most of the time. To test this thing place a url 
`https://こんにち` in the end of line in bash and start moving it through the end by char and highlight it. You should see that chars sometimes are not fully underlined or not moving to the next line, etc, etc. I guess selection across multiple lines not always working for double width chars is kind of related to that.