@chrisduerr 2019-12-16
What do you mean by `moving it` and how does this behave before this patch?

I can't reproduce it, though I'm not sure I understand what you mean.