@kchibisov 2019-12-16
The performance of this branch is questionable btw.
```
Master:
nightly asrw 6.00107 +- 0.00357 seconds time elapsed  ( +-  0.06% )
stable  asrw 6.36960 +- 0.00118 seconds time elapsed  ( +-  0.02% )

PR:
nightly asrw 6.35584 +- 0.00284 seconds time elapsed  ( +-  0.04% )
stable asrw 5.84472 +- 0.00150 seconds time elapsed  ( +-  0.03% )
```

I mean it feels like there's a "flip" in rustc, since there's something similar between this benchmark results. I'm not sure what we should do about it, probably nothing, but ideally we can figure out what is causing this results and tune to be the same as our master. And I'm sure that I was benchmarking the right terminals.