@kchibisov 2019-12-16
@chrisduerr I hope these gifs can help you with my comment.

PR: 
![bad](https://user-images.githubusercontent.com/27620401/70912268-a35e5000-2024-11ea-8280-05667fb0b82b.gif)

master:
![good](https://user-images.githubusercontent.com/27620401/70912316-bcff9780-2024-11ea-8a4c-24af60b7de10.gif)