@chrisduerr 2019-12-16
> The performance of this branch is questionable btw.

Which nightly are you running?

> @chrisduerr I hope these gifs can help you with my comment.

Thanks, I wasn't aware this is about URL highlighting.