@chrisduerr 2019-12-16
@kchibisov I still can't reproduce the bash issue with URLs btw. It is also noteworthy that both ZSH and bash seem to do their own custom shenanigans when double-width chars are in the last column.