@kchibisov 2019-12-16
The 2 bugs I've mentioned were the bugs related to my old (bash-4.4.23), the bash 5.0.11 is fine (at least on my 2 examples ;- )).