@chrisduerr 2019-12-23
@kchibisov reflow should now behave correctly, feel free to give that another shot if you have some time.

I think the functionality should be in a good place now, though I have to test everything again just to make sure. I still need to clean up the grow_cols and shrink_cols methods since they're pretty massive now, but that shouldn't affect functionality.