@chrisduerr 2019-12-24
I wasn't really able to clean up the `shrink_cols` and `grow_cols` methods much, since there's a lot to them and moving things into other functions quickly leads to massive lists of parameters with issues about multiple mutable borrowers.

Likely the only way to clean this up significantly would be to rewrite it, but I don't want to make such big changes in this PR and I'm not sure if that would improve anything. So I think it should be fine for now hopefully.

I've made some fixes to the behavior of shrinking column which should now hopefully give us fully accurate shrinking and growing behavior. I've also copied VTE's behavior of not performing this kind of reflow in the alt screen, which I think makes sense.